+++ src/app/(mypage)/wish-list/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from 'next/navigation';
 import { createClient } from '@/lib/supabase/server';
-import type { Artwork } from '@/components/myArtworks/Types';
-import WishlistScreen from '@/components/wishList/WishlistScreen';
+import type { Artwork } from '@/components/my-artworks/Types';
+import WishlistScreen from '@/components/my-wishlist/WishlistScreen';
 
 type LikeRow = {
   artworks: {
